requirements.txt
numpy
Pillow
torchvision
tqdm
onnx==1.8.1
onnxruntime==1.6.0
opencv-python==4.5.1.48
torch==1.7.1
Flask==2.0.3
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.39
PyJWT==2.6.0
gunicorn==20.0.4
bcrypt==4.0.1
python-dotenv==0.21.1
psycopg2-binary==2.9.5